requirements.txt
streamlit==1.39.0
pandas==2.2.2
numpy==2.1.2
plotly==5.24.1
selenium==4.26.1
webdriver-manager==4.0.2
openpyxl==3.1.5
python-dotenv==1.0.1
